# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Order Hub API — multi-platform inventory and order management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/revenue:
    get:
      operationId: getDashboardRevenue
      tags: [dashboard]
      summary: Daily revenue series for the last N days
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            minimum: 7
            maximum: 90
            default: 30
      responses:
        "200":
          description: Revenue series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyRevenuePoint"
  /dashboard/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [dashboard]
      summary: Most recent orders for the activity panel
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies (brands)
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict (slug already exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /companies/{slug}:
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company (admin only). Fails if it still has data.
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Cannot delete (last company or has associated data)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/top-skus:
    get:
      operationId: getTopSkus
      tags: [dashboard]
      summary: Top SKUs by revenue
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
      responses:
        "200":
          description: Top SKUs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopSku"
...
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a single inventory item by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory quantity / price / product name
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Platform:
      type: string
      enum: [meesho, flipkart, amazon]
    Brand:
      type: string
      description: |
        Active brand workspace identifier (a company `slug`). Sent on every
        authenticated request via the `X-Brand` header. Companies are managed
        dynamically via the `/companies` endpoints.
    Company:
      type: object
      properties:
        id:
          type: integer
        slug:
          $ref: "#/components/schemas/Brand"
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, slug, name, createdAt]
    CompanyInput:
      type: object
      properties:
        slug:
          type: string
          description: URL-safe identifier (1-32 chars, lowercase a-z 0-9 and dashes).
        name:
          type: string
          minLength: 1
          maxLength: 80
      required: [slug, name]
...
    DailyRevenuePoint:
      type: object
      properties:
        date:
          type: string
          format: date
        revenue:
          type: number
        orders:
          type: integer
      required: [date, revenue, orders]
    TopSku:
      type: object
      properties:
        sku:
          type: string
        productName:
          type: string
        platform:
          $ref: "#/components/schemas/Platform"
        units:
          type: integer
        revenue:
          type: number
      required: [sku, productName, platform, units, revenue]
    Order:
      type: object
      properties:
        id:
          type: integer
        brand:
          $ref: "#/components/schemas/Brand"
        platform:
          $ref: "#/components/schemas/Platform"
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        price:
          type: number
          description: Unit price
        orderDate:
          type: string
          format: date-time
        subOrderNo:
          type: ["string", "null"]
          description: Marketplace sub order number (e.g. Meesho Sub Order No)
        orderSource:
          type: ["string", "null"]
          description: Marketplace order source (e.g. "Ad order")
        reasonForCredit:
          type: ["string", "null"]
          description: Order status / Reason for Credit Entry (e.g. DELIVERED, SHIPPED, CANCELLED, RTO_LOCKED)
        createdAt:
          type: string
          format: date-time
      required: [id, brand, platform, productName, sku, quantity, price, orderDate, createdAt]
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        brand:
          $ref: "#/components/schemas/Brand"
        platform:
          $ref: "#/components/schemas/Platform"
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        gstPercent:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        purchaseDate:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, brand, platform, productName, sku, quantity, price, updatedAt]
...
    InventoryUpdate:
      type: object
      properties:
        platform:
          $ref: "#/components/schemas/Platform"
        productName:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 0
        price:
          type: number
          minimum: 0
        originalPrice:
          type: number
          minimum: 0
        gstPercent:
          type: number
          minimum: 0
          maximum: 100
        imageUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        purchaseDate:
          type: string
          format: date-time
...